diff_service/requirements.txt
flask
redis
deepdiff
jsonschema
pyyaml
prometheus-client
prometheus-flask-exporter